feat: (SOLE/hperf#17) update requirements file and README.md
Now the requirements to run hperf is simplified.
We change the requirements from 'requirements.txt' to 'environment.yml',
which currently only contains 'pandas' and 'numpy' and specifies the version of python.

B.t.w., the README.md is updated since the way to create environment is changed.
The English version of README, 'README.en.md', is consistent with the Chinese version.