feat: (SOLE/hperf#15) add sanity check
We add method 'sanity check' in 'Profiler' to conduct sanity check.
'Controller' is responsible for calling this method before profiling.

Besides, we need to identify the ISA of the SUT when we check whether NMI watchdog is enabled,
so we modified the methods in 'EventGroup' and its member 'isa' can be accessed in 'Profiler'.