feat: (SOLE/hperf#21) use 'Logger' instead of 'logging'
use logging.getLogger() to get an instance of 'Logger' and use its method to log, instead of using logging.info(), etc.

set the log output to both console and file,
the Logger will be firstly initialized so that the log file will be saved in /tmp/hperf/hperf.log and then copied to the test directory.

a method __save_log_file() is added in 'Controller' and it will be used at the end of this running.

in detail, this commit is to solve the problem metioned in issue #21:
https://git.solelab.tech/SOLE/hperf/issues/21#issuecomment-181