fix: exit hperf when run profiling script failed on SUT
'Profiler' did not check the returned code of profiling script running on the SUT,
so that if there is any error encountered, no raw performance data will be generated and 'Analyzer' will fail to read.

To fix the problem, if the returned code is not equal to 0, hperf will output the error message and terminate hperf.